feat: add chat panel with loading skeleton to reduce js bundle
