+++ src/components/Chat.tsx
@@ -1,123 +1,76 @@
 import { useChatbot } from "@/contexts/ChatContext";
-import { useChat } from "ai/react";
-import { useCallback, useEffect, useRef, type ChangeEvent } from "react";
+import { Suspense, lazy, useCallback, useState } from "react";
 import ChatHeader from "./ChatHeader";
-import ChatInput from "./ChatInput";
-import ChatMessages from "./ChatMessages";
 import {
   Accordion,
   AccordionContent,
   AccordionItem,
   AccordionTrigger,
 } from "./ui/Accordion";
+import { Skeleton } from "./ui/skeleton";
+
+const ChatPanel = lazy(() => import("./ChatPanel"));
+
+function ChatPanelFallback() {
+  return (
+    <div className="flex flex-1 flex-col justify-between">
+      <div className="flex flex-1 flex-col justify-end gap-3 overflow-hidden p-2 sm:gap-4 sm:p-3">
+        <div className="flex items-start justify-end">
+          <Skeleton className="h-10 w-[220px] rounded-lg sm:w-64" />
+        </div>
+
+        <div className="flex items-start justify-start">
+          <Skeleton className="mr-2 mt-0.5 h-4 w-4 shrink-0 rounded-full sm:mr-2.5 sm:h-5 sm:w-5" />
+          <Skeleton className="h-20 w-[220px] rounded-lg sm:w-64" />
+        </div>
+
+        <div className="flex items-start justify-end">
+          <Skeleton className="h-10 w-[220px] rounded-lg sm:w-64" />
+        </div>
+
+        <div className="flex items-start justify-start">
+          <Skeleton className="mr-2 mt-0.5 h-4 w-4 shrink-0 rounded-full sm:mr-2.5 sm:h-5 sm:w-5" />
+          <Skeleton className="h-20 w-[220px] rounded-lg sm:w-64" />
+        </div>
+      </div>
+
+      <div className="flex gap-1.5 border-t px-2 py-2 backdrop-blur-sm sm:gap-2 sm:px-3 sm:py-2.5">
+        <Skeleton className="h-9 w-10 sm:h-10 sm:w-12" />
+        <Skeleton className="h-8 flex-1 sm:h-9" />
+        <Skeleton className="h-9 w-10 sm:h-10 sm:w-12" />
+      </div>
+
+      <span className="sr-only" role="status" aria-live="polite">
+        Loading chat…
+      </span>
+    </div>
+  );
+}
 
 export default function Chat() {
-  const chatIdRef = useRef<string | null>(null);
-
-  const chatFetch = useCallback<typeof fetch>(
-    async (input: RequestInfo | URL, init?: RequestInit) => {
-      const baseInit: RequestInit = init || {};
-      const headers = new Headers(baseInit.headers);
-      const originalBody = baseInit.body;
-      let bodyToSend = originalBody;
-
-      const addSessionToPayload = (chatId: string | null) => {
-        if (!chatId) return;
-        headers.set("X-Chat-Id", chatId);
-
-        if (typeof bodyToSend === "string") {
-          try {
-            const parsedBody = JSON.parse(bodyToSend);
-            parsedBody.chat_id = chatId;
-            bodyToSend = JSON.stringify(parsedBody);
-          } catch (error) {
-            console.error("[chat] failed to attach chat_id", error);
-          }
-        }
-      };
-
-      addSessionToPayload(chatIdRef.current);
-
-      const performFetch = async (
-        headersOverride: HeadersInit,
-        bodyOverride: BodyInit | null | undefined,
-      ) =>
-        fetch(input, {
-          ...baseInit,
-          headers: headersOverride,
-          body: bodyOverride,
-        });
-
-      const captureSessionId = (res: Response) => {
-        const newChatId = res.headers.get("X-Chat-Id");
-        if (newChatId) {
-          chatIdRef.current = newChatId;
-        }
-      };
-
-      let response = await performFetch(headers, bodyToSend);
-
-      if (response.status === 400) {
-        const errorText = await response
-          .clone()
-          .text()
-          .catch(() => "");
-
-        if (errorText.includes("Invalid X-Chat-Id")) {
-          chatIdRef.current = null;
-
-          const retryHeaders = new Headers(baseInit.headers);
-          retryHeaders.delete("X-Chat-Id");
-          retryHeaders.delete("x-chat-id");
-          let retryBody = originalBody;
-
-          if (typeof originalBody === "string") {
-            try {
-              const parsedRetryBody = JSON.parse(originalBody);
-              delete parsedRetryBody.chat_id;
-              retryBody = JSON.stringify(parsedRetryBody);
-            } catch (error) {
-              console.error("[chat] failed to strip chat_id", error);
-            }
-          }
-
-          response = await performFetch(retryHeaders, retryBody);
-          captureSessionId(response);
-          return response;
-        }
-      }
-
-      captureSessionId(response);
-      return response;
-    },
-    [],
-  );
-
-  const {
-    messages,
-    input,
-    handleInputChange,
-    handleSubmit,
-    setMessages,
-    isLoading,
-    error,
-  } = useChat({ fetch: chatFetch });
-
-  useEffect(() => {
-    if (messages.length === 0) {
-      chatIdRef.current = null;
-    }
-  }, [messages]);
-
   const { isVisible } = useChatbot();
 
-  const handleClearChat = () => {
-    chatIdRef.current = null;
-  };
+  const [expandedValue, setExpandedValue] = useState<string>("");
+  const [hasOpened, setHasOpened] = useState(false);
+
+  const handleValueChange = useCallback((nextValue: string) => {
+    setExpandedValue(nextValue);
+    if (nextValue) {
+      setHasOpened(true);
+    }
+  }, []);
+
+  const isExpanded = expandedValue === "item-1";
 
   return (
     isVisible && (
-      <Accordion type="single" collapsible className="relative z-40 flex">
+      <Accordion
+        type="single"
+        collapsible
+        value={expandedValue}
+        onValueChange={handleValueChange}
+        className="relative z-40 flex"
+      >
         <AccordionItem
           value="item-1"
           className="fixed bottom-4 right-4 w-[320px] rounded-lg border bg-background shadow-lg shadow-black/10 sm:bottom-8 sm:right-8 sm:w-96 dark:shadow-black/30"
@@ -125,26 +78,23 @@ export default function Chat() {
           <AccordionTrigger className="border-b px-6">
             <ChatHeader />
           </AccordionTrigger>
-          <AccordionContent className="flex max-h-[400px] min-h-[350px] flex-col justify-between rounded-b-lg p-0 sm:max-h-[500px] sm:min-h-[400px]">
-            <ChatMessages
-              messages={messages}
-              error={error}
-              isLoading={isLoading}
-              onPromptClick={(prompt) =>
-                handleInputChange({
-                  target: { value: prompt },
-                } as ChangeEvent<HTMLInputElement>)
-              }
-            />
-            <ChatInput
-              input={input}
-              handleSubmit={handleSubmit}
-              handleInputChange={handleInputChange}
-              setMessages={setMessages}
-              onClearChat={handleClearChat}
-              isLoading={isLoading}
-              messages={messages}
-            />
+          <AccordionContent
+            forceMount={hasOpened ? true : undefined}
+            className="p-0"
+          >
+            {hasOpened && (
+              <div
+                className={
+                  isExpanded
+                    ? "flex max-h-[400px] min-h-[350px] flex-col justify-between rounded-b-lg sm:max-h-[500px] sm:min-h-[400px]"
+                    : "hidden"
+                }
+              >
+                <Suspense fallback={<ChatPanelFallback />}>
+                  <ChatPanel isExpanded={isExpanded} />
+                </Suspense>
+              </div>
+            )}
           </AccordionContent>
         </AccordionItem>
       </Accordion>
